+++ plugin/src/main/java/dev/flikas/spring/boot/assistant/idea/plugin/inspection/yaml/PropertyDeprecatedInspection.java
@@ -1,4 +1,4 @@
-package dev.flikas.spring.boot.assistant.idea.plugin.inspection;
+package dev.flikas.spring.boot.assistant.idea.plugin.inspection.yaml;
 
 import com.intellij.codeInspection.ProblemsHolder;
 import dev.flikas.spring.boot.assistant.idea.plugin.metadata.index.MetadataProperty;