+++ plugin/src/main/java/dev/flikas/spring/boot/assistant/idea/plugin/inspection/yaml/YamlInspectionBase.java
@@ -1,4 +1,4 @@
-package dev.flikas.spring.boot.assistant.idea.plugin.inspection;
+package dev.flikas.spring.boot.assistant.idea.plugin.inspection.yaml;
 
 import com.intellij.codeInspection.LocalInspectionTool;
 import com.intellij.codeInspection.LocalInspectionToolSession;