+++ plugin/src/main/java/icons/Icons.java
@@ -1,0 +1,9 @@
+package icons;
+
+import javax.swing.*;
+
+import static com.intellij.openapi.util.IconLoader.getIcon;
+
+public class Icons {
+  public static final Icon SpringBoot = getIcon("/icons/springBoot.svg", Icons.class);
+}